requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dict==2020.12.3
easy==0.0.4
easydict==1.13
enum34==1.1.10
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.1
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.7
Jinja2==2.7.3
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prettytensor==0.7.4
progressbar==2.5
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==3.11
requests==2.32.2
rich==13.7.1
sh==2.0.6
six==1.16.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
torch==2.3.0
torchfile==0.1.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
values==2020.12.3
Werkzeug==3.0.3
wrapt==1.16.0
